Get rid of the iconv dependency
The iconv package does not work properly on modern macOS producing empty
output for "ASCII" to "UTF8" conversion. Also, with last Hackage upload
in 2015 and no source repository specified it seems abandoned. Given
that it was already an inconenience on Windows it makes sense to ditch
this dependency altogether.

The `text` package provides decoding facilities for UTF-8 and UTF-16
encoding and we already depend on that package. There is a trade-off
though: we cede support for JIS enconding. If anybody needs it back and
can provide an example file with it, we can implement it using the
`encoding` package later.

https://www.awaresystems.be/imaging/tiff/tifftags/privateifd/exif/usercomment.html